Refactor particle management out
