+++ cifront/app/scripts/app.js
@@ -1,40 +1,14 @@
 import "../styles/index.scss";
-import { debounce, throttle, defer, sample } from "lodash";
+import { debounce, defer, sample, throttle } from "lodash";
 import $ from "jquery";
 import ace from "brace";
 import "brace/mode/html";
 import "brace/theme/vibrant_ink";
 import "brace/ext/searchbox";
+import { ParticleHandler } from "./particles";
 
 const POWER_MODE_ACTIVATION_THRESHOLD = 200;
 const STREAK_TIMEOUT = 10 * 1000;
-
-const MAX_PARTICLES = 500;
-const PARTICLE_NUM_RANGE = [5, 6, 7, 8, 9, 10, 11, 12];
-const PARTICLE_GRAVITY = 0.075;
-const PARTICLE_SIZE = 8;
-const PARTICLE_ALPHA_FADEOUT = 0.96;
-const PARTICLE_VELOCITY_RANGE = {
-  x: [-2.5, 2.5],
-  y: [-7, -3.5]
-};
-
-const PARTICLE_COLORS = {
-  text: [255, 255, 255],
-  "text.xml": [255, 255, 255],
-  keyword: [0, 221, 255],
-  variable: [0, 221, 255],
-  "meta.tag.tag-name.xml": [0, 221, 255],
-  "keyword.operator.attribute-equals.xml": [0, 221, 255],
-  constant: [249, 255, 0],
-  "constant.numeric": [249, 255, 0],
-  "support.constant": [249, 255, 0],
-  "string.attribute-value.xml": [249, 255, 0],
-  "string.unquoted.attribute-value.html": [249, 255, 0],
-  "entity.other.attribute-name.xml": [129, 148, 244],
-  comment: [0, 255, 121],
-  "comment.xml": [0, 255, 121]
-};
 
 const EXCLAMATION_EVERY = 10;
 const EXCLAMATIONS = [
@@ -59,8 +33,6 @@ class App {
   constructor() {
     this.currentStreak = 0;
     this.powerMode = false;
-    this.particles = [];
-    this.particlePointer = 0;
     this.lastDraw = 0;
 
     this.$streakCounter = $(".streak-container .counter");
@@ -70,11 +42,10 @@ class App {
     this.$nameTag = $(".name-tag");
     this.$result = $(".result");
     this.$editor = $("#editor");
-    this.canvas = this.setupCanvas();
-    this.canvasContext = this.canvas.getContext("2d");
     this.$finish = $(".finish-button");
-
     this.$body = $("body");
+
+    this.particleHandler = new ParticleHandler($(".canvas-overlay")[0]);
 
     this.debouncedSaveContent = debounce(this.saveContent, 300);
     this.debouncedEndStreak = debounce(this.endStreak, STREAK_TIMEOUT);
@@ -120,13 +91,6 @@ class App {
     return editor;
   }
 
-  setupCanvas() {
-    const canvas = $(".canvas-overlay")[0];
-    canvas.width = window.innerWidth;
-    canvas.height = window.innerHeight;
-    return canvas;
-  }
-
   getName(forceUpdate) {
     let { name } = localStorage;
     if (!name || forceUpdate) name = prompt("What's your name?");
@@ -151,9 +115,10 @@ class App {
   };
 
   onFrame = time => {
-    this.drawParticles(time - this.lastDraw);
+    this.particleHandler.drawParticles(time - this.lastDraw);
     this.lastDraw = time;
-    if (window.requestAnimationFrame) window.requestAnimationFrame(this.onFrame);
+    if (window.requestAnimationFrame)
+      window.requestAnimationFrame(this.onFrame);
   };
 
   increaseStreak() {
@@ -226,67 +191,9 @@ class App {
     if (!this.powerMode) {
       return;
     }
-
     const { x, y } = this.getCursorPosition();
-    const numParticles = sample(PARTICLE_NUM_RANGE);
-    const color = this.getParticleColor(type);
-    for (let i = 0; i < numParticles; i++) {
-      this.particles[this.particlePointer] = this.createParticle(x, y, color);
-      this.particlePointer = (this.particlePointer + 1) % MAX_PARTICLES;
-    }
+    this.particleHandler.spawnParticles(x, y, type);
   }
-
-  getParticleColor(type) {
-    return PARTICLE_COLORS[type] || [255, 255, 255];
-  }
-
-  createParticle(x, y, color) {
-    return {
-      x,
-      y: y + 10,
-      alpha: 1,
-      color,
-      velocity: {
-        x:
-          PARTICLE_VELOCITY_RANGE.x[0] +
-          Math.random() *
-            (PARTICLE_VELOCITY_RANGE.x[1] - PARTICLE_VELOCITY_RANGE.x[0]),
-        y:
-          PARTICLE_VELOCITY_RANGE.y[0] +
-          Math.random() *
-            (PARTICLE_VELOCITY_RANGE.y[1] - PARTICLE_VELOCITY_RANGE.y[0])
-      }
-    };
-  }
-
-  drawParticles = () => {
-    this.canvasContext.clearRect(0, 0, this.canvas.width, this.canvas.height);
-
-    const result = [];
-    this.particles.forEach(particle => {
-      if (particle.alpha <= 0.1) {
-        return;
-      }
-
-      particle.velocity.y += PARTICLE_GRAVITY;
-      particle.x += particle.velocity.x;
-      particle.y += particle.velocity.y;
-      particle.alpha *= PARTICLE_ALPHA_FADEOUT;
-
-      this.canvasContext.fillStyle = `rgba(${particle.color.join(", ")}, ${
-        particle.alpha
-      })`;
-      result.push(
-        this.canvasContext.fillRect(
-          Math.round(particle.x - PARTICLE_SIZE / 2),
-          Math.round(particle.y - PARTICLE_SIZE / 2),
-          PARTICLE_SIZE,
-          PARTICLE_SIZE
-        )
-      );
-    });
-    return result;
-  };
 
   shake() {
     if (!this.powerMode) {
