+++ package.json
@@ -8,7 +8,7 @@
     "dev": "webpack --mode=development --watch --progress",
     "build": "webpack --mode=production --progress",
     "lint": "eslint cifront/app/scripts",
-    "prettify": "prettier --write cifront/app/scripts/app.js"
+    "prettify": "prettier --write cifront/app/scripts/*.js"
   },
   "dependencies": {
     "brace": "^0.11.1",
